# docs/swagger.yaml
basePath: /
definitions:
...
  api.ConfirmSlackCodeRequest:
    properties:
      code:
        type: string
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.RequestSlackCodeRequest:
    properties:
      slack_user_id:
        type: string
    type: object
...
  model.ExternalIdentity:
    properties:
      chat_id:
        description: provider-specific (Telegram DM chat)
        type: string
      created_at:
        type: string
      display_name:
        description: shown in timeline / UI
        type: string
      external_id:
        type: string
      id:
        type: string
      provider:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  model.User:
    properties:
      auth_provider:
        description: '"password" or "oidc" - last login method'
        type: string
      created_at:
        type: string
      email:
        description: Unique email
        type: string
      id:
        description: UUID
        type: string
      identities:
        description: loaded on demand by handlers
        items:
          $ref: '#/definitions/model.ExternalIdentity'
        type: array
      name:
        description: Display name
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.UserRole'
        description: 'Global role: admin or user'
    type: object
  model.UserRole:
    enum:
    - admin
    - user
    type: string
    x-enum-comments:
      UserRoleAdmin: Full system access
      UserRoleUser: Standard authenticated user (default)
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleUser
...
info:
  contact: {}
  description: TokayOps API
  title: TokayOps API
  version: "1.0"
paths:
...
  /api/auth/me/slack:
    delete:
      consumes:
      - application/json
      description: Remove the Slack identity binding from the current user
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Unbind Slack account
      tags:
      - auth
  /api/auth/me/slack/confirm-code:
    post:
      consumes:
      - application/json
      description: Verify the OTP code and bind the Slack account
      parameters:
      - description: Confirm data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ConfirmSlackCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Confirm Slack OTP code
      tags:
      - auth
  /api/auth/me/slack/request-code:
    post:
      consumes:
      - application/json
      description: Send an OTP code to the specified Slack User ID via DM
      parameters:
      - description: Request data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.RequestSlackCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Request a Slack OTP code
      tags:
      - auth
  /api/auth/me/telegram:
    delete:
      description: Remove the Telegram identity binding from the current user
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Unbind Telegram account
      tags:
      - auth
  /api/auth/me/telegram/link:
    post:
      description: Issue a one-time deep-link token and return a t.me/<bot>?start=<token>
        URL. The user opens it and presses Start to link their Telegram account.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Request a Telegram deep link
      tags:
      - auth
...
swagger: "2.0"